Remove SslKeyStoreConfigurator/SslTrustStoreConfigurator concept
