+++ config-model/src/test/java/com/yahoo/vespa/model/container/xml/JettyContainerModelBuilderTest.java
@@ -3,30 +3,21 @@ package com.yahoo.vespa.model.container.xml;
 
 import com.yahoo.config.model.builder.xml.test.DomBuilderTest;
 import com.yahoo.container.ComponentsConfig;
-import com.yahoo.container.bundle.BundleInstantiationSpecification;
 import com.yahoo.container.jdisc.FilterBindingsProvider;
 import com.yahoo.jdisc.http.ConnectorConfig;
-import com.yahoo.jdisc.http.ssl.DefaultSslKeyStoreConfigurator;
-import com.yahoo.jdisc.http.ssl.DefaultSslTrustStoreConfigurator;
 import com.yahoo.vespa.model.container.ContainerCluster;
 import com.yahoo.vespa.model.container.component.SimpleComponent;
 import com.yahoo.vespa.model.container.http.ConnectorFactory;
 import com.yahoo.vespa.model.container.http.JettyHttpServer;
-
 import com.yahoo.vespa.model.container.http.ssl.DefaultSslProvider;
 import org.junit.Test;
 import org.w3c.dom.Element;
-import org.xml.sax.SAXException;
 
-import java.io.IOException;
-import java.util.Arrays;
 import java.util.List;
 import java.util.Optional;
-import java.util.Set;
 
 import static com.yahoo.jdisc.http.ConnectorConfig.Ssl.KeyStoreType;
 import static org.hamcrest.CoreMatchers.equalTo;
-import static org.hamcrest.CoreMatchers.hasItem;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.CoreMatchers.not;
 import static org.hamcrest.CoreMatchers.nullValue;
@@ -197,44 +188,6 @@ public class JettyContainerModelBuilderTest extends ContainerModelBuilderTestBas
     }
 
     @Test
-    public void ssl_keystore_and_truststore_configurator_can_be_overriden() throws IOException, SAXException {
-        Element clusterElem = DomBuilderTest.parse(
-                "<jdisc id='default' version='1.0' jetty='true'>",
-                "  <http>",
-                "    <server port='9000' id='foo'>",
-                "      <ssl-keystore-configurator class='com.yahoo.MySslKeyStoreConfigurator' bundle='mybundle'/>",
-                "      <ssl-truststore-configurator class='com.yahoo.MySslTrustStoreConfigurator' bundle='mybundle'/>",
-                "    </server>",
-                "    <server port='9001' id='bar'/>",
-                "  </http>",
-                nodesXml,
-                "</jdisc>");
-        createModel(root, clusterElem);
-        ContainerCluster cluster = (ContainerCluster) root.getChildren().get("default");
-        List<ConnectorFactory> connectorFactories = cluster.getChildrenByTypeRecursive(ConnectorFactory.class);
-        {
-            ConnectorFactory firstConnector = connectorFactories.get(0);
-            assertConnectorHasInjectedComponents(firstConnector, "ssl-keystore-configurator@foo", "ssl-truststore-configurator@foo", "dummy-ssl-provider@foo");
-            assertComponentHasClassNameAndBundle(getChildComponent(firstConnector, 0),
-                                                 "com.yahoo.MySslKeyStoreConfigurator",
-                                                 "mybundle");
-            assertComponentHasClassNameAndBundle(getChildComponent(firstConnector, 1),
-                                                 "com.yahoo.MySslTrustStoreConfigurator",
-                                                 "mybundle");
-        }
-        {
-            ConnectorFactory secondConnector = connectorFactories.get(1);
-            assertConnectorHasInjectedComponents(secondConnector, "ssl-keystore-configurator@bar", "ssl-truststore-configurator@bar", "dummy-ssl-provider@bar");
-            assertComponentHasClassNameAndBundle(getChildComponent(secondConnector, 0),
-                                                 DefaultSslKeyStoreConfigurator.class.getName(),
-                                                 "jdisc_http_service");
-            assertComponentHasClassNameAndBundle(getChildComponent(secondConnector, 1),
-                                                 DefaultSslTrustStoreConfigurator.class.getName(),
-                                                 "jdisc_http_service");
-        }
-    }
-
-    @Test
     public void verify_that_ssl_element_generates_connector_config_and_inject_provider_component() {
         Element clusterElem = DomBuilderTest.parse(
                 "<jdisc id='default' version='1.0' jetty='true'>",
@@ -313,25 +266,6 @@ public class JettyContainerModelBuilderTest extends ContainerModelBuilderTestBas
         List<ConnectorFactory> connectorFactories = cluster.getChildrenByTypeRecursive(ConnectorFactory.class);
         ConnectorFactory connectorFactory = connectorFactories.get(0);
         assertChildComponentExists(connectorFactory, "com.yahoo.CustomSslProvider");
-    }
-
-    private static void assertConnectorHasInjectedComponents(ConnectorFactory connectorFactory, String... componentNames) {
-        Set<String> injectedComponentIds = connectorFactory.getInjectedComponentIds();
-        assertThat(injectedComponentIds.size(), equalTo(componentNames.length));
-        Arrays.stream(componentNames)
-                .forEach(name -> assertThat(injectedComponentIds, hasItem(name)));
-    }
-
-    private static SimpleComponent getChildComponent(ConnectorFactory connectorFactory, int index) {
-        return connectorFactory.getChildrenByTypeRecursive(SimpleComponent.class).get(index);
-    }
-
-    private static void assertComponentHasClassNameAndBundle(SimpleComponent simpleComponent,
-                                                             String className,
-                                                             String bundleName) {
-        BundleInstantiationSpecification spec = simpleComponent.model.bundleInstantiationSpec;
-        assertThat(spec.classId.toString(), is(className));
-        assertThat(spec.bundle.toString(), is(bundleName));
     }
 
     private static void assertChildComponentExists(ConnectorFactory connectorFactory, String className) {
