+++ config-model/src/test/schema-test-files/services.xml
@@ -112,10 +112,7 @@
         </request-chain>
       </filtering>
 
-      <server port="4080" id="myServer">
-        <ssl-keystore-configurator class="com.yahoo.MySslKeyStoreConfigurator" bundle="mybundle" />
-        <ssl-truststore-configurator class="com.yahoo.MySslTrustStoreConfigurator" bundle="mybundle" />
-      </server>
+      <server port="4080" id="myServer"/>
       <server port="4081" id="anotherServer">
         <config name="container.jdisc.config.http-server">
           <maxChunkSize>9999</maxChunkSize>
